SPU-3962

+++ lib/features/buy_flow/ui/buy_tab_body.dart
@@ -114,8 +114,9 @@ class _BuyAmountScreenBodyState extends State<BuyAmountTabBody> with AutomaticKe
                     pasteLabel: intl.paste,
                     onPaste: () async {
                       final data = await Clipboard.getData('text/plain');
+
                       if (data?.text != null) {
-                        final n = int.tryParse(data!.text!);
+                        final n = double.tryParse(data!.text!);
                         if (n != null) {
                           store.pasteValue(n.toString().trim());
                         }
